+++ cmd/server/main.go
@@ -22,7 +22,7 @@ import (
 )
 
 const (
-	Version = "3.3.1"
+	Version = "3.3.2"
 )
 
 var (